world generation doesn't crash, still need to check distribution (#580)

+++ extrabees/src/main/java/binnie/extrabees/utils/config/ConfigurationMain.java
@@ -6,28 +6,28 @@ public class ConfigurationMain implements IConfigurable {
 
 	private static final String WORLDGEN = "World-Gen";
 	private static boolean canQuarryMineHives = true;
-	private static int waterHiveRate = 2;
-	private static int rockHiveRate = 2;
-	private static int netherHiveRate = 2;
-	private static int marbleHiveRate = 2;
+	private static float waterHiveRate = 2.0F;
+	private static float rockHiveRate = 2.0F;
+	private static float netherHiveRate = 2.0F;
+	private static float marbleHiveRate = 2.0F;
 
 	public static boolean isCanQuarryMineHives() {
 		return canQuarryMineHives;
 	}
 
-	public static int getWaterHiveRate() {
+	public static float getWaterHiveRate() {
 		return waterHiveRate;
 	}
 
-	public static int getRockHiveRate() {
+	public static float getRockHiveRate() {
 		return rockHiveRate;
 	}
 
-	public static int getNetherHiveRate() {
+	public static float getNetherHiveRate() {
 		return netherHiveRate;
 	}
 
-	public static int getMarbleHiveRate() {
+	public static float getMarbleHiveRate() {
 		return marbleHiveRate;
 	}
 
@@ -36,9 +36,9 @@ public class ConfigurationMain implements IConfigurable {
 		canQuarryMineHives = config.getBoolean("canQuarryMineHives", Configuration.CATEGORY_GENERAL, true, "Sets whether a quarry will be able to mine hives or not.");
 
 		config.addCustomCategoryComment(WORLDGEN, "WorldGen settings for ExtraBees");
-		waterHiveRate = config.getInt("waterHiveRate", WORLDGEN, 2, 0, 10, "Sets the worldgen spawn chance for water hives.");
-		rockHiveRate = config.getInt("rockHiveRate", WORLDGEN, 2, 0, 10, "Sets the worldgen spawn chance for rock hives.");
-		netherHiveRate = config.getInt("netherHiveRate", WORLDGEN, 2, 0, 10, "Sets the worldgen spawn chance for nether hives.");
-		marbleHiveRate = config.getInt("marbleHiveRate", WORLDGEN, 2, 0, 10, "Sets the worldgen spawn chance for marble hives.");
+		waterHiveRate = config.getFloat("waterHiveRate", WORLDGEN, 2, 0, 10, "Sets the worldgen spawn chance for water hives.");
+		rockHiveRate = config.getFloat("rockHiveRate", WORLDGEN, 2, 0, 10, "Sets the worldgen spawn chance for rock hives.");
+		netherHiveRate = config.getFloat("netherHiveRate", WORLDGEN, 2, 0, 10, "Sets the worldgen spawn chance for nether hives.");
+		marbleHiveRate = config.getFloat("marbleHiveRate", WORLDGEN, 2, 0, 10, "Sets the worldgen spawn chance for marble hives.");
 	}
 }
